Fix hardcoded group size in buffer tests
Request of work items adjusts to the available capabilities of underlying
device in buffer tests.

Function to query available amount of work items moved from
tests/group_functions/group_functions_common.h to
tests/common/get_group_range.h.
