+++ tests/group_functions/group_of.h
@@ -44,7 +44,7 @@ void joint_of_group(sycl::queue& queue) {
   const std::string test_cases_names[test_cases] = {"none true", "one true",
                                                     "some true", "all true"};
 
-  sycl::range<D> work_group_range = util::work_group_range<D>(queue);
+  sycl::range<D> work_group_range = sycl_cts::util::work_group_range<D>(queue);
   size_t work_group_size = work_group_range.size();
 
   const size_t sizes[3] = {5, work_group_size / 2, 3 * work_group_size};
@@ -162,7 +162,8 @@ void joint_of_group(sycl::queue& queue) {
     int index = 0;
     for (int i = 0; i < test_matrix; ++i)
       for (int j = 0; j < test_cases; ++j) {
-        std::string work_group = util::work_group_print(work_group_range);
+        std::string work_group =
+            sycl_cts::util::work_group_print(work_group_range);
         CAPTURE(D, work_group);
         INFO("Value of " << test_names[i] << " with " << test_cases_names[j]
                          << " predicate"
@@ -193,7 +194,7 @@ void predicate_function_of_group(sycl::queue& queue) {
   const std::string test_cases_names[test_cases] = {"none true", "one true",
                                                     "some true", "all true"};
 
-  sycl::range<D> work_group_range = util::work_group_range<D>(queue);
+  sycl::range<D> work_group_range = sycl_cts::util::work_group_range<D>(queue);
   size_t work_group_size = work_group_range.size();
 
   // array to return results: 4 predicates * 3 functions
@@ -251,7 +252,8 @@ void predicate_function_of_group(sycl::queue& queue) {
   int index = 0;
   for (int i = 0; i < test_matrix; ++i)
     for (int j = 0; j < test_cases; ++j) {
-      std::string work_group = util::work_group_print(work_group_range);
+      std::string work_group =
+          sycl_cts::util::work_group_print(work_group_range);
       CAPTURE(D, work_group);
       INFO("Value of " << test_names[i] << " with " << test_cases_names[j]
                        << " predicate is " << (res[index] ? "right" : "wrong"));
@@ -280,7 +282,7 @@ void predicate_function_of_sub_group(sycl::queue& queue) {
   const std::string test_cases_names[test_cases] = {"none true", "one true",
                                                     "some true", "all true"};
 
-  sycl::range<D> work_group_range = util::work_group_range<D>(queue);
+  sycl::range<D> work_group_range = sycl_cts::util::work_group_range<D>(queue);
 
   // array to return results: 4 predicates * 3 functions
   bool res[test_matrix * test_cases] = {false};
@@ -338,7 +340,8 @@ void predicate_function_of_sub_group(sycl::queue& queue) {
   int index = 0;
   for (int i = 0; i < test_matrix; ++i)
     for (int j = 0; j < test_cases; ++j) {
-      std::string work_group = util::work_group_print(work_group_range);
+      std::string work_group =
+          sycl_cts::util::work_group_print(work_group_range);
       CAPTURE(D, work_group);
       INFO("Value of " << test_names[i] << " with " << test_cases_names[j]
                        << " predicate is " << (res[index] ? "right" : "wrong"));
@@ -367,7 +370,7 @@ void bool_function_of_group(sycl::queue& queue) {
 
   using T = size_t;
 
-  sycl::range<D> work_group_range = util::work_group_range<D>(queue);
+  sycl::range<D> work_group_range = sycl_cts::util::work_group_range<D>(queue);
   size_t work_group_size = work_group_range.size();
 
   // array to return results: 4 predicates * 3 functions
@@ -423,7 +426,8 @@ void bool_function_of_group(sycl::queue& queue) {
   int index = 0;
   for (int i = 0; i < test_matrix; ++i)
     for (int j = 0; j < test_cases; ++j) {
-      std::string work_group = util::work_group_print(work_group_range);
+      std::string work_group =
+          sycl_cts::util::work_group_print(work_group_range);
       CAPTURE(D, work_group);
       INFO("Value of " << test_names[i] << " with " << test_cases_names[j]
                        << " predicate is " << (res[index] ? "right" : "wrong"));
@@ -452,7 +456,7 @@ void bool_function_of_sub_group(sycl::queue& queue) {
 
   using T = size_t;
 
-  sycl::range<D> work_group_range = util::work_group_range<D>(queue);
+  sycl::range<D> work_group_range = sycl_cts::util::work_group_range<D>(queue);
 
   // array to return results
   bool res[test_matrix * test_cases] = {false};
@@ -511,7 +515,8 @@ void bool_function_of_sub_group(sycl::queue& queue) {
   int index = 0;
   for (int i = 0; i < test_matrix; ++i)
     for (int j = 0; j < test_cases; ++j) {
-      std::string work_group = util::work_group_print(work_group_range);
+      std::string work_group =
+          sycl_cts::util::work_group_print(work_group_range);
       CAPTURE(D, work_group);
       INFO("Value of " << test_names[i] << " with " << test_cases_names[j]
                        << " predicate is " << (res[index] ? "right" : "wrong"));
