+++ tests/group_functions/group_scan.h
@@ -47,7 +47,7 @@ void joint_scan_group(sycl::queue& queue) {
   constexpr int test_cases = 2;
   const std::string test_cases_names[test_cases] = {"plus", "maximum"};
 
-  sycl::range<D> work_group_range = util::work_group_range<D>(queue);
+  sycl::range<D> work_group_range = sycl_cts::util::work_group_range<D>(queue);
   size_t work_group_size = work_group_range.size();
 
   const size_t sizes[3] = {5, work_group_size / 2, 3 * work_group_size};
@@ -243,7 +243,8 @@ void joint_scan_group(sycl::queue& queue) {
     int index = 0;
     for (int i = 0; i < test_matrix; ++i)
       for (int j = 0; j < test_cases; ++j) {
-        std::string work_group = util::work_group_print(work_group_range);
+        std::string work_group =
+            sycl_cts::util::work_group_print(work_group_range);
         CAPTURE(D, work_group, size);
         INFO("Value of " << test_names[i] << " with " << test_cases_names[j]
                          << " operation"
@@ -298,7 +299,7 @@ void init_joint_scan_group(sycl::queue& queue) {
   constexpr int test_cases = 2;
   const std::string test_cases_names[test_cases] = {"plus", "maximum"};
 
-  sycl::range<D> work_group_range = util::work_group_range<D>(queue);
+  sycl::range<D> work_group_range = sycl_cts::util::work_group_range<D>(queue);
   size_t work_group_size = work_group_range.size();
 
   const size_t sizes[3] = {5, work_group_size / 2, 3 * work_group_size};
@@ -479,7 +480,8 @@ void init_joint_scan_group(sycl::queue& queue) {
     int index = 0;
     for (int i = 0; i < test_matrix; ++i)
       for (int j = 0; j < test_cases; ++j) {
-        std::string work_group = util::work_group_print(work_group_range);
+        std::string work_group =
+            sycl_cts::util::work_group_print(work_group_range);
         CAPTURE(D, work_group, size);
         INFO("Value of " << test_names[i] << " with " << test_cases_names[j]
                          << " operation"
@@ -535,7 +537,7 @@ void scan_over_group(sycl::queue& queue) {
   constexpr int test_cases = 2;
   const std::string test_cases_names[test_cases] = {"plus", "maximum"};
 
-  sycl::range<D> work_group_range = util::work_group_range<D>(queue);
+  sycl::range<D> work_group_range = sycl_cts::util::work_group_range<D>(queue);
   size_t work_group_size = work_group_range.size();
 
   // array to return results:
@@ -660,7 +662,8 @@ void scan_over_group(sycl::queue& queue) {
       for (size_t k = 1; k < work_group_size; ++k)
         result &= res[index * work_group_size + k];
 
-      std::string work_group = util::work_group_print(work_group_range);
+      std::string work_group =
+          sycl_cts::util::work_group_print(work_group_range);
       CAPTURE(D, work_group);
       INFO("Value of " << test_names[i] << " with " << test_cases_names[j]
                        << " operation"
@@ -722,7 +725,7 @@ void init_scan_over_group(sycl::queue& queue) {
   constexpr int test_cases = 2;
   const std::string test_cases_names[test_cases] = {"plus", "maximum"};
 
-  sycl::range<D> work_group_range = util::work_group_range<D>(queue);
+  sycl::range<D> work_group_range = sycl_cts::util::work_group_range<D>(queue);
   size_t work_group_size = work_group_range.size();
 
   // array to return results:
@@ -868,7 +871,8 @@ void init_scan_over_group(sycl::queue& queue) {
       for (size_t k = 1; k < work_group_size; ++k)
         result &= res[index * work_group_size + k];
 
-      std::string work_group = util::work_group_print(work_group_range);
+      std::string work_group =
+          sycl_cts::util::work_group_print(work_group_range);
       CAPTURE(D, work_group);
       INFO("Value of " << test_names[i] << " with " << test_cases_names[j]
                        << " operation"
